Dirty programming hack
With this code it is possible to partially program the fan.  Work to decode
the initial header required, it is possible that there is some compression
happening as the loop to send the image doesn't replace an existing image.
The opening byte sequence after the initial header is always 0x4023.